Update to Batch File Issue Line 73
The output file is now able to be moved with the edits to lines 71-73, however the file is not able to be copied to the root directory at this stage. Will need to check and make sure this does work on the windows 10 VM before closing this issue.

Also discovered a bug on line 63 which now outputs the hidden files of the root directory to the screen.